Separate main.rs into main.rs and lib.rs
Keeping main.rs slim and doing most logic in the library.
* Separates CLI stuff from the functionality of the program.
* Makes it possible to have doc tests in lib.rs
* Can potentially make the library more testable
